+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0
+
+* Filter domain records with argument -t and the type (mx/cname...).
+* List domains with "gopapi domains".
+
 # 1.0.2
 
 * Improved authentication cipher security.